corr

+++ src/core/losses.py
@@ -73,7 +73,7 @@ def accuracy(inputs,
         acc = (inputs == receiver_output)  # [batch_size,n_attributes]
 
         if all_attributes_equal:
-            acc = 1 * th.all(acc, dim=1)  # [batch_size]
+            acc = (1 * th.all(acc, dim=1)).float()  # [batch_size]
         else:
             acc = (1 * acc).float().mean(dim=1)  # [batch_size]
